F#17 - WIP on persistent dropdowns
