+++ server/upload/uploads/Products.csv
@@ -1,0 +1,6 @@
+product_id,prouct_name,Description,color,price,user_id
+1,A,aa,aaa,5,1
+2,B,bb,bbb,10,2
+3,C,cc,ccc,15,3
+4,D,dd,ddd,20,4
+5,E,cc,eee,10,5
